requirements.txt
streamlit==1.44.1
streamlit-feedback==0.1.4
mistralai==0.4.2
langchain_community==0.3.27
faiss-cpu==1.10.0
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.1.0
requests==2.32.3
